Move some common functions to specific helper files
