+++ Sources/SwiftGenKit/Parsers/FontsParser/CTFont.swift
@@ -9,20 +9,6 @@ import AppKit.NSFont
 import Foundation
 import PathKit
 
-extension PathKit.Path {
-  /// Returns the Path relative to a parent directory.
-  /// If the argument passed as parent isn't a prefix of `self`, returns `nil`.
-  ///
-  /// - Parameter parent: The parent Path to get the relative path against
-  /// - Returns: The relative Path, or nil if parent was not a parent dir of self
-  func relative(to parent: Path) -> Path? {
-    let parentComponents = parent.absolute().components
-    let currentComponents = self.absolute().components
-    return currentComponents.starts(with: parentComponents)
-      ? Path(components: currentComponents.dropFirst(parentComponents.count)) : nil
-  }
-}
-
 extension CTFont {
   static func parse(file: Path, relativeTo parent: Path? = nil) -> [Font] {
     let descs = CTFontManagerCreateFontDescriptorsFromURL(file.url as CFURL) as NSArray?
